feat: moved dark mode button to its own component
