+++ tsconfig.json
@@ -38,7 +38,7 @@
     }
   },
   "include": [
-    "src",
+    "src/**/*",
     "next-env.d.ts"
   ],
   "exclude": [
